split index.js into multiple files
